optimized tmux_window_fire.sh
tmux_window_fire.sh is way too slow to be run by tmux multiple times
(one for each window) in a blocking way.
This is a workaround: now the script updates the status every 5s and
takes advantage of the fact that tmux supports getting updates from a
continuously running process in #().
But it is still too slow and causes many instances of the script to be
running in parallel. A neater solution needs to be found.